+++ frontend/src/hooks/Groups/useGroupsData.tsx
@@ -1,15 +1,6 @@
 import { useGetGroupsQuery } from "../../graphql/getGroups.graphql.types";
+import { Group, SearchStudent } from "../../utils";
 import { useEditionSelection } from "../common/useEditionSelection";
-
-export type GroupData = {
-  name: string;
-  id: string;
-};
-
-export type SearchStudentData = {
-  fullName?: string;
-  userId: string;
-};
 
 export const useGroupsData = () => {
   const { selectedEdition } = useEditionSelection();
@@ -20,22 +11,20 @@ export const useGroupsData = () => {
     },
   });
 
-  const groups: GroupData[] | undefined = data?.edition[0].groups.map(
-    (group) => {
-      return {
-        name: group.groupName,
-        id: group.groupsId,
-      };
-    },
-  );
+  const groups: Group[] | undefined = data?.edition[0].groups.map((group) => {
+    return {
+      name: group.groupName,
+      id: group.groupsId,
+    };
+  });
 
-  const students: SearchStudentData[] | undefined =
-    data?.edition[0].groups.flatMap((group) =>
+  const students: SearchStudent[] | undefined = data?.edition[0].groups.flatMap(
+    (group) =>
       group.userGroups.map((userGroup) => ({
         fullName: userGroup.user.fullName ?? undefined,
         userId: userGroup.user.userId,
       })),
-    );
+  );
 
   return { loading, error, groups, students };
 };
