--- frontend/src/screens/Groups/SearchStudent/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchStudent";
