+++ frontend/src/screens/Groups/StudentSearcher/StudentCard.tsx
