+++ frontend/src/screens/Groups/StudentSearcher/index.ts
@@ -1,0 +1,1 @@
+export * from "./StudentSearcher";
